Use "go vet" instead of deprecated "go vet tool"
As of Go 1.12, `go vet tool` has been deprecated.

See: https://golang.org/doc/go1.12#vet

Support for the `-shadow` flag has been dropped as it was experimental
and now requires an external tool in order to run. If it is desirable
to include, then an option to configure the `-vettool` flag would be
more useful.